refactoring classes for model
