+++ code/aind_smartspim_classification/model/__init__.py
@@ -1,0 +1,3 @@
+"""
+Custom model package for cell detection
+"""
