+++ code/run_capsule.py
@@ -145,12 +145,12 @@ def set_up_pipeline_parameters(pipeline_config: dict, default_config: dict):
         Dictionary with the combined parameters
     """
 
-    default_config[
-        "input_channel"
-    ] = f"{pipeline_config['segmentation']['channel']}.zarr"
-    default_config[
-        "background_channel"
-    ] = f"{pipeline_config['segmentation']['background_channel']}.zarr"
+    default_config["input_channel"] = (
+        f"{pipeline_config['segmentation']['channel']}.zarr"
+    )
+    default_config["background_channel"] = (
+        f"{pipeline_config['segmentation']['background_channel']}.zarr"
+    )
     default_config["channel"] = pipeline_config["segmentation"]["channel"]
     default_config["input_scale"] = pipeline_config["segmentation"]["input_scale"]
     default_config["chunk_size"] = int(128)
